Split up the tests some
