+++ baiji/test_s3.py
@@ -1,91 +1,12 @@
-# FIXME pylint: disable=protected-access, unnecessary-lambda, unused-variable
 import os
 import shutil
 import unittest
-import uuid
 import mock
 from baiji import s3
 from baiji.util.testing import create_random_temporary_file
 from baiji.util import tempfile
 
 TEST_BUCKET = 'baiji-test'
-
-class TestS3TmpDir(unittest.TestCase):
-    '''
-    We test this separately since we use it in TestAWSBase.setUp()
-
-    '''
-    def setUp(self):
-        self.s3_path = 'tmp/' + str(uuid.uuid4()) + '/'
-
-    def tearDown(self):
-        b = s3.S3Connection()._bucket(TEST_BUCKET)
-        for key in b.list(self.s3_path[0:-1]):
-            b.delete_key(key)
-
-    def test_s3_tmpdir(self):
-        def fake_uuid():
-            fake_uuid.counter += 1
-            return "FAKE-UUID-%d" % fake_uuid.counter
-        fake_uuid.counter = 0
-        self.assertEqual(
-            s3.path.gettmpdir(bucket=TEST_BUCKET, prefix=self.s3_path, uuid_generator=fake_uuid),
-            's3://%s/%sFAKE-UUID-1/' % (TEST_BUCKET, self.s3_path))
-        self.assertEqual(
-            len(list(s3.ls('s3://%s/%sFAKE-UUID-1' % (TEST_BUCKET, self.s3_path)))),
-            1)
-        self.assertTrue(
-            s3.exists('s3://%s/%sFAKE-UUID-1/.tempdir' % (TEST_BUCKET, self.s3_path)))
-        self.assertEqual(
-            s3.path.gettmpdir(bucket=TEST_BUCKET, prefix=self.s3_path, uuid_generator=fake_uuid),
-            's3://%s/%sFAKE-UUID-2/' % (TEST_BUCKET, self.s3_path))
-        self.assertEqual(len(list(s3.ls('s3://%s/%sFAKE-UUID-2' % (TEST_BUCKET, self.s3_path)))), 1)
-        self.assertTrue(s3.exists('s3://%s/%sFAKE-UUID-2/.tempdir' % (TEST_BUCKET, self.s3_path)))
-
-class TestS3Path(unittest.TestCase):
-
-    def test_s3_parse(self):
-        def test(s, expect, on_windows=None):
-            if on_windows and os.name == 'nt':
-                expect = expect[:2] + (on_windows,)
-            o = s3.path.parse(s)
-            for ii in range(3):
-                self.assertEqual(o[ii], expect[ii])
-
-        test('s3://foo/path/to/bar', ('s3', 'foo', '/path/to/bar'))
-        test('s3://xx', ('s3', 'xx', ''))
-        test('s3://xx/', ('s3', 'xx', '/'))
-        self.assertRaises(ValueError, s3.path.parse, 's3:///path/to/bar')
-        test('file:///path/to/blah', ('file', '', '/path/to/blah'), on_windows=r'C:\path\to\blah')
-        test('/path/to/blah', ('file', '', '/path/to/blah'), on_windows=r'C:\path\to\blah')
-        test('path/to/blah', ('file', '', os.path.abspath('./path/to/blah')))
-        test('~/path/to/blah', ('file', '', os.path.abspath(os.path.expanduser('~/path/to/blah'))))
-        test('', ('file', '', os.path.abspath('.')))
-
-        self.assertTrue(s3.path.islocal('/path/to/bar'))
-        self.assertFalse(s3.path.islocal('s3://foo/path/to/bar'))
-        self.assertFalse(s3.path.isremote('/path/to/bar'))
-        self.assertTrue(s3.path.isremote('s3://foo/path/to/bar'))
-
-        self.assertEqual(s3.path.join('s3://foo/path', 'to/bar'), 's3://foo/path/to/bar')
-        if os.name == 'nt':
-            self.assertEqual(s3.path.join('/foo/path', 'to/bar'), r'C:\foo\path\to\bar')
-        else:
-            self.assertEqual(s3.path.join('/foo/path', 'to/bar'), '/foo/path/to/bar')
-
-        self.assertEqual(s3.path.basename('file:///foo/path/to/bar.baz'), 'bar.baz')
-        self.assertEqual(s3.path.basename('s3://bucket/foo/path/to/bar.baz'), 'bar.baz')
-
-        test('C:\\path\\of\\quux', ('file', '', 'C:\\path\\of\\quux'))
-        test('file:///C:\\path\\of\\quux', ('file', '', 'C:\\path\\of\\quux'))
-
-        with self.assertRaises(ValueError):
-            s3.path.parse('file:///foo;bar')
-
-        # Make sure it preserves trailing slashes
-        test('path/to/blah/', ('file', '', os.path.join(os.path.abspath('./path/to/blah'), '')))
-        test('s3://foo/path/to/blah/', ('s3', 'foo', '/path/to/blah/'))
-
 
 class TestAWSBase(unittest.TestCase):
     """
@@ -104,7 +25,7 @@ class TestAWSBase(unittest.TestCase):
         loc = s3.path.parse(self.s3_test_location)
         self.s3_path = loc.path
 
-        b = s3.S3Connection()._bucket(self.bucket)
+        b = s3.S3Connection()._bucket(self.bucket) # pylint: disable=protected-access
         for key in b.list(self.s3_path[1:]):
             b.delete_key(key)
         self.assertEqual(len([x for x in b.list(self.s3_path[1:])]), 0) # just make sure we're starting with a clean slate
@@ -116,7 +37,7 @@ class TestAWSBase(unittest.TestCase):
     def tearDown(self):
         shutil.rmtree(self.tmp_dir, ignore_errors=True)
         os.remove(self.local_file)
-        b = s3.S3Connection()._bucket(self.bucket)
+        b = s3.S3Connection()._bucket(self.bucket) # pylint: disable=protected-access
         for key in b.list(self.s3_path[1:]):
             b.delete_key(key)
         self.assertEqual(len([x for x in b.list(self.s3_path[1:])]), 0)
@@ -142,7 +63,7 @@ class TestAWSBase(unittest.TestCase):
     def assert_is_public(self, s3_url, is_public):
         from urlparse import urlparse
         url = urlparse(s3_url)
-        acl = self.retriable_s3_call(lambda: s3.S3Connection()._bucket(url.netloc).get_acl(url.path[1:]))
+        acl = self.retriable_s3_call(lambda: s3.S3Connection()._bucket(url.netloc).get_acl(url.path[1:])) # pylint: disable=protected-access
         actual_is_public = False
         for grant in acl.acl.grants:
             if grant.permission == 'READ':
@@ -163,20 +84,6 @@ class TestAWSBase(unittest.TestCase):
         if not s3.exists(uri):
             s3.cp(self.local_file, uri)
         return uri
-
-
-class TestAWS(TestAWSBase):
-
-    def test_credentials(self):
-        from baiji.config import credentials
-        from baiji.util import yaml
-        bodylabs_file_path = os.getenv('BODYLABS_CREDENTIAL_FILE', os.path.expanduser('~/.bodylabs'))
-        if not os.path.exists(bodylabs_file_path):
-            raise unittest.SkipTest("Skipping test_credentials because ~/.bodylabs doesn't exist.")
-        truth = yaml.load(bodylabs_file_path)
-        self.assertEqual(credentials.key, truth['AWS_ACCESS_KEY'])
-        self.assertEqual(credentials.secret, truth['AWS_SECRET'])
-
 
 class TestS3Exists(TestAWSBase):
 
@@ -205,7 +112,6 @@ class TestS3Exists(TestAWSBase):
 
     @mock.patch('baiji.s3.S3Connection._lookup')
     def test_s3_exists_return_false_if_the_file_never_shows_up(self, mock_lookup):
-        mock_key = "all_we_care_is_that_it's not None"
         mock_lookup.return_value = None
         self.assertFalse(s3.exists('s3://foo'))
         self.assertEqual(mock_lookup.call_count, 3)
@@ -422,8 +328,8 @@ class TestS3(TestAWSBase):
         for f in files:
             s3.cp(self.local_file, self.remote_file(f))
         self.assertEqual(set(s3.ls(self.s3_test_location)), set(map(lambda x: self.s3_path+x, files)))
-        self.assertEqual(set(s3.ls(self.s3_test_location, return_full_urls=True)), set(map(lambda x: self.remote_file(x), files)))
-        self.assertEqual(set(s3.ls(self.s3_test_location, shallow=True)), set(map(lambda x: self.s3_path+x, ['foo', 'bar.baz', 'quack/'])))
+        self.assertEqual(set(s3.ls(self.s3_test_location, return_full_urls=True)), set([self.remote_file(x) for x in files]))
+        self.assertEqual(set(s3.ls(self.s3_test_location, shallow=True)), set([self.s3_path+x for x in ['foo', 'bar.baz', 'quack/']]))
 
     def test_s3_md5(self):
         s3.cp(self.local_file, self.remote_file("file_to_md5"))
@@ -443,7 +349,7 @@ class TestS3(TestAWSBase):
             s3.cp(self.local_file, self.remote_file(f))
         glob_results = list(s3.glob(self.s3_test_location, '*.obj'))
         self.assertEqual(2, len(glob_results))
-        self.assertEqual(set(glob_results), set(map(lambda f: self.remote_file(f), ['a.obj', 'b.obj'])))
+        self.assertEqual(set(glob_results), set([self.remote_file(f) for f in ['a.obj', 'b.obj']]))
 
     def test_s3_glob_match_multiple_wildcards(self):
         files = ['body_1_pose_T.obj', 'body_1_pose_Fun.obj', 'body_2_pose_T.obj', 'body_02_pose_T.obj']
@@ -453,8 +359,7 @@ class TestS3(TestAWSBase):
         self.assertEqual(3, len(glob_results))
         self.assertEqual(
             set(glob_results),
-            set(map(lambda f: self.remote_file(f),
-                    ['body_1_pose_T.obj', 'body_1_pose_Fun.obj', 'body_2_pose_T.obj']))
+            set([self.remote_file(f) for f in ['body_1_pose_T.obj', 'body_1_pose_Fun.obj', 'body_2_pose_T.obj']])
         )
 
     def test_s3_glob_exclude(self):
@@ -463,7 +368,7 @@ class TestS3(TestAWSBase):
             s3.cp(self.local_file, self.remote_file(f))
         glob_results = list(s3.glob(self.s3_test_location, 'pose_[!T].obj'))
         self.assertEqual(1, len(glob_results))
-        self.assertEqual(set(glob_results), set(map(lambda f: self.remote_file(f), ['pose_A.obj'])))
+        self.assertEqual(set(glob_results), set([self.remote_file('pose_A.obj')]))
 
     def test_s3_with_double_slashes_in_key(self):
         '''
@@ -523,468 +428,9 @@ class TestEncryption(TestAWSBase):
         self.assertTrue(s3.info(self.remote_file("encrypted.txt"))['encrypted'])
 
 
-class TestSync(TestAWSBase):
-    def setUp(self):
-        super(TestSync, self).setUp()
-
-        # Files that are the same on both sides:
-        self.create_random_file_at((self.tmp_dir, self.s3_test_location), 'foo/a.txt')
-        self.create_random_file_at((self.tmp_dir, self.s3_test_location), 'foo/b.txt')
-        self.create_random_file_at((self.tmp_dir, self.s3_test_location), 'foo/bar/c.txt')
-        self.create_random_file_at((self.tmp_dir, self.s3_test_location), 'foo/bar/baz/d.txt')
-        # Files that are different locally and remotely:
-        self.create_random_file_at((self.tmp_dir,), 'foo/x.txt')
-        self.create_random_file_at((self.s3_test_location,), 'foo/x.txt')
-        self.create_random_file_at((self.tmp_dir,), 'foo/bar/x.txt')
-        self.create_random_file_at((self.s3_test_location,), 'foo/bar/x.txt')
-        # Files that are only local:
-        self.create_random_file_at((self.tmp_dir,), 'foo/loc.txt')
-        self.create_random_file_at((self.tmp_dir,), 'foo/bar/loc.txt')
-        # Files that are only remote:
-        self.create_random_file_at((self.s3_test_location,), 'foo/rem.txt')
-        self.create_random_file_at((self.s3_test_location,), 'foo/bar/rem.txt')
-
-        self.local_dir_to_sync = s3.path.join(self.tmp_dir, 'foo')
-        self.remote_dir_to_sync = s3.path.join(self.s3_test_location, 'foo')
-
-        self.expected_local_contents = s3.ls(self.local_dir_to_sync)
-        self.expected_remote_contents = [x.replace(self.s3_path+'foo/', '') for x in s3.ls(self.remote_dir_to_sync)]
-
-    def create_random_file_at(self, bases, path):
-        from baiji.util.testing import random_data
-        data = random_data()
-        for base in bases:
-            with s3.open(s3.path.join(base, path), 'w') as f:
-                f.write(data)
-
-    def assertContentsAre(self, expected_contents):
-        self.assertSetEqual(set(s3.ls(self.local_dir_to_sync)), set(expected_contents))
-        self.assertSetEqual(set([x.replace(self.s3_path+'foo/', '') for x in s3.ls(self.remote_dir_to_sync)]), set(expected_contents))
-        # TODO files are equal:
-
-
-    def test_sync_local_to_remote(self):
-        s3.sync(self.local_dir_to_sync, self.remote_dir_to_sync)
-        self.assertContentsAre(self.expected_local_contents)
-
-    def test_sync_remote_to_local(self):
-        s3.sync(self.remote_dir_to_sync, self.local_dir_to_sync)
-        self.assertContentsAre(self.expected_remote_contents)
-
-    @mock.patch('baiji.s3.S3Connection.cp')
-    @mock.patch('baiji.s3.S3Connection.rm')
-    def test_sync_file_same(self, rm, cp):
-        # In these tests, we want to check that rm and cp are invoked only
-        # when they should be, so we mock out cp and rm on a new instance of
-        # S3Connection. (It seems difficult to use mock.patch on module
-        # methods.)
-        #
-        # We also need s3.path.join but it's defined on the module, not on
-        # instances of S3Connection.
-        from baiji.s3 import S3Connection
-
-        s3_with_mocks = S3Connection()
-
-        s3_with_mocks.sync_file(
-            s3.path.join(self.local_dir_to_sync, 'a.txt'),
-            s3.path.join(self.remote_dir_to_sync, 'a.txt')
-        )
-
-        self.assertFalse(cp.called)
-        self.assertFalse(rm.called)
-
-    @mock.patch('baiji.s3.S3Connection.cp')
-    @mock.patch('baiji.s3.S3Connection.rm')
-    def test_sync_file_only_src(self, rm, cp):
-        from baiji.s3 import S3Connection
-
-        s3_with_mocks = S3Connection()
-
-        s3_with_mocks.sync_file(
-            s3.path.join(self.local_dir_to_sync, 'loc.txt'),
-            s3.path.join(self.remote_dir_to_sync, 'loc.txt')
-        )
-
-        cp.assert_called_once_with(
-            s3.path.join(self.local_dir_to_sync, 'loc.txt'),
-            s3.path.join(self.remote_dir_to_sync, 'loc.txt'),
-            force=False,
-            progress=False,
-            policy=None,
-            encoding=None,
-            encrypt=True,
-            guess_content_type=False
-        )
-        self.assertFalse(rm.called)
-
-    @mock.patch('baiji.s3.S3Connection.cp')
-    @mock.patch('baiji.s3.S3Connection.rm')
-    def test_sync_file_only_dst(self, rm, cp):
-        from baiji.s3 import S3Connection
-
-        s3_with_mocks = S3Connection()
-
-        # Test with delete=True and delete=False.
-
-        s3_with_mocks.sync_file(
-            s3.path.join(self.local_dir_to_sync, 'rem.txt'),
-            s3.path.join(self.remote_dir_to_sync, 'rem.txt'),
-            delete=True
-        )
-
-        self.assertFalse(cp.called)
-        rm.assert_called_once_with(
-            s3.path.join(self.remote_dir_to_sync, 'rem.txt')
-        )
-
-        rm.reset_mock()
-
-        with self.assertRaises(s3.KeyNotFound):
-            s3_with_mocks.sync_file(
-                s3.path.join(self.local_dir_to_sync, 'rem.txt'),
-                s3.path.join(self.remote_dir_to_sync, 'rem.txt'),
-                delete=False
-            )
-
-        self.assertFalse(cp.called)
-        self.assertFalse(rm.called)
-
-    @mock.patch('baiji.s3.S3Connection.cp')
-    @mock.patch('baiji.s3.S3Connection.rm')
-    def test_sync_file_exists_but_outdated(self, rm, cp):
-        from baiji.s3 import S3Connection
-
-        s3_with_mocks = S3Connection()
-
-        # Test with update=True and update=False.
-
-        s3_with_mocks.sync_file(
-            s3.path.join(self.local_dir_to_sync, 'x.txt'),
-            s3.path.join(self.remote_dir_to_sync, 'x.txt'),
-            update=True
-        )
-
-        cp.assert_called_once_with(
-            s3.path.join(self.local_dir_to_sync, 'x.txt'),
-            s3.path.join(self.remote_dir_to_sync, 'x.txt'),
-            force=True,
-            progress=False,
-            policy=None,
-            encoding=None,
-            encrypt=True,
-            guess_content_type=False
-        )
-        self.assertFalse(rm.called)
-
-        cp.reset_mock()
-
-        s3_with_mocks.sync_file(
-            s3.path.join(self.local_dir_to_sync, 'x.txt'),
-            s3.path.join(self.remote_dir_to_sync, 'x.txt'),
-            update=False
-        )
-
-        self.assertFalse(cp.called)
-        self.assertFalse(rm.called)
-
-
-class TestCachedFile(TestAWSBase):
-    def setUp(self):
-        super(TestCachedFile, self).setUp()
-        s3.cp(self.local_file, self.remote_file("openable"))
-        with open(self.local_file, 'rb') as f:
-            self.truth = f.read()
-
-    def test_s3_open_read_remote_file_with_context_manager(self):
-        self.assert_s3_exists(self.remote_file("openable"))
-
-        with s3.open(self.remote_file("openable"), 'r') as f:
-            tempname = f.name
-            self.assertEqual(self.truth, f.read())
-
-        self.assertFalse(os.path.exists(tempname))
-
-    def test_s3_open_read_remote_file_without_context_manager(self):
-        self.assert_s3_exists(self.remote_file("openable"))
-
-        f = s3.open(self.remote_file("openable"), 'r')
-        tempname = f.name
-        self.assertEqual(self.truth, f.read())
-        f.close()
-
-        self.assertFalse(os.path.exists(tempname))
-
-    def test_s3_open_read_local_file_with_context_manager(self):
-        self.assert_s3_exists(self.local_file)
-
-        with s3.open(self.local_file, 'r') as f:
-            self.assertEqual(self.truth, f.read())
-            self.assertEqual(f.name, self.local_file)
-
-    def test_s3_open_read_local_file_without_context_manager(self):
-        self.assert_s3_exists(self.local_file)
-
-        f = s3.open(self.local_file, 'r')
-        self.assertEqual(self.truth, f.read())
-        self.assertEqual(f.name, self.local_file)
-        f.close()
-
-    def test_s3_open_write_remote_file_with_context_manager(self):
-        remote_file_name = self.remote_file("write_test_1")
-        local_file_name = os.path.join(self.tmp_dir, "write_test_1")
-        self.assert_s3_does_not_exist(remote_file_name)
-
-        with s3.open(remote_file_name, 'w') as f:
-            tempname = f.name
-            f.write(self.truth)
-
-        self.assertFalse(os.path.exists(tempname))
-        self.assert_s3_exists(remote_file_name)
-        # download and confirm that it contains the correct contents
-        s3.cp(remote_file_name, local_file_name)
-        with open(local_file_name) as f:
-            self.assertEqual(self.truth, f.read())
-
-    def test_s3_open_write_local_file_creates_parent_dirs(self):
-        local_file_name = os.path.join(self.tmp_dir, "subdir", "write_test_subdir")
-        self.assert_s3_does_not_exist(local_file_name)
-        self.assert_s3_does_not_exist(os.path.dirname(local_file_name))
-
-        with s3.open(local_file_name, 'w') as f:
-            f.write(self.truth)
-
-        self.assert_s3_exists(local_file_name)
-        self.assert_s3_exists(os.path.dirname(local_file_name))
-
-    def test_s3_open_write_remote_file_without_context_manager(self):
-        remote_file_name = self.remote_file("write_test_2")
-        local_file_name = os.path.join(self.tmp_dir, "write_test_2")
-        self.assert_s3_does_not_exist(remote_file_name)
-
-        f = s3.open(remote_file_name, 'w')
-        tempname = f.name
-        f.write(self.truth)
-        f.close()
-
-        self.assertFalse(os.path.exists(tempname))
-        self.assert_s3_exists(remote_file_name)
-        # download and confirm that it contains the correct contents
-        s3.cp(remote_file_name, local_file_name)
-        with open(local_file_name) as f:
-            self.assertEqual(self.truth, f.read())
-
-    def test_s3_open_write_local_file_with_context_manager(self):
-        local_file_name = os.path.join(self.tmp_dir, "write_test_local_1")
-        self.assert_s3_does_not_exist(local_file_name)
-
-        with s3.open(local_file_name, 'w') as f:
-            self.assertEqual(f.name, local_file_name)
-            f.write(self.truth)
-
-        self.assert_s3_exists(local_file_name)
-        # confirm that it contains the correct contents
-        with open(local_file_name) as f:
-            self.assertEqual(self.truth, f.read())
-
-    def test_s3_open_write_local_file_without_context_manager(self):
-        local_file_name = os.path.join(self.tmp_dir, "write_test_local_2")
-        self.assert_s3_does_not_exist(local_file_name)
-
-        f = s3.open(local_file_name, 'w')
-        self.assertEqual(f.name, local_file_name)
-        f.write(self.truth)
-        f.close()
-
-        self.assert_s3_exists(local_file_name)
-        # confirm that it contains the correct contents
-        with open(local_file_name) as f:
-            self.assertEqual(self.truth, f.read())
-
-    def test_s3_open_exclusive_write_remote_file_throws_error_for_existing_file(self):
-        self.assert_s3_exists(self.remote_file("openable"))
-        with self.assertRaises(s3.KeyExists):
-            s3.open(self.remote_file("openable"), 'x')
-
-    def test_s3_open_exclusive_write_local_file_throws_error_for_existing_file(self):
-        self.assert_s3_exists(self.local_file)
-        with self.assertRaises(s3.KeyExists):
-            s3.open(self.local_file, 'x')
-
-    def test_s3_open_exclusive_write_remote_file_throws_error_for_double_open(self):
-        self.assert_s3_does_not_exist(self.remote_file("exclusive_write_test_3"))
-        with s3.open(self.remote_file("exclusive_write_test_3"), 'x') as f1:
-            with self.assertRaises(s3.KeyExists):
-                with s3.open(self.remote_file("exclusive_write_test_3"), 'x') as f2:
-                    pass
-
-    def test_s3_open_exclusive_write_local_file_throws_error_for_double_open(self):
-        local_file_name = os.path.join(self.tmp_dir, "exclusive_write_test_local_3")
-        self.assert_s3_does_not_exist(local_file_name)
-        with s3.open(local_file_name, 'x') as f1:
-            with self.assertRaises(s3.KeyExists):
-                with s3.open(local_file_name, 'x') as f2:
-                    pass
-
-    def test_s3_open_exclusive_write_remote_file_with_context_manager(self):
-        remote_file_name = self.remote_file("exclusive_write_test_1")
-        local_file_name = os.path.join(self.tmp_dir, "exclusive_write_test_1")
-        self.assert_s3_does_not_exist(remote_file_name)
-
-        with s3.open(remote_file_name, 'x') as f:
-            self.assert_s3_exists(remote_file_name)
-            tempname = f.name
-            f.write(self.truth)
-
-        self.assertFalse(os.path.exists(tempname))
-        self.assert_s3_exists(remote_file_name)
-        # download and confirm that it contains the correct contents
-        s3.cp(remote_file_name, local_file_name)
-        with open(local_file_name) as f:
-            self.assertEqual(self.truth, f.read())
-
-    def test_s3_open_exclusive_write_remote_file_without_context_manager(self):
-        remote_file_name = self.remote_file("exclusive_write_test_2")
-        local_file_name = os.path.join(self.tmp_dir, "exclusive_write_test_2")
-        self.assert_s3_does_not_exist(remote_file_name)
-
-        f = s3.open(remote_file_name, 'x')
-        self.assert_s3_exists(remote_file_name)
-        tempname = f.name
-        f.write(self.truth)
-        f.close()
-
-        self.assertFalse(os.path.exists(tempname))
-        self.assert_s3_exists(remote_file_name)
-        # download and confirm that it contains the correct contents
-        s3.cp(remote_file_name, local_file_name)
-        with open(local_file_name) as f:
-            self.assertEqual(self.truth, f.read())
-
-    def test_s3_open_exclusive_write_local_file_with_context_manager(self):
-        local_file_name = os.path.join(self.tmp_dir, "exclusive_write_test_local_1")
-        self.assert_s3_does_not_exist(local_file_name)
-
-        with s3.open(local_file_name, 'x') as f:
-            self.assertEqual(f.name, local_file_name)
-            f.write(self.truth)
-
-        self.assert_s3_exists(local_file_name)
-        # confirm that it contains the correct contents
-        with open(local_file_name) as f:
-            self.assertEqual(self.truth, f.read())
-
-    def test_s3_open_exclusive_write_local_file_without_context_manager(self):
-        local_file_name = os.path.join(self.tmp_dir, "exclusive_write_test_local_2")
-        self.assert_s3_does_not_exist(local_file_name)
-
-        f = s3.open(local_file_name, 'x')
-        self.assertEqual(f.name, local_file_name)
-        f.write(self.truth)
-        f.close()
-
-        self.assert_s3_exists(local_file_name)
-        # confirm that it contains the correct contents
-        with open(local_file_name) as f:
-            self.assertEqual(self.truth, f.read())
-
-    def test_s3_open_read_raises_error_for_nonexistent_remote_file(self):
-        nonexistent_file = self.remote_file(str(uuid.uuid4()))
-        self.assert_s3_does_not_exist(nonexistent_file)
-        with self.assertRaises(s3.KeyNotFound):
-            s3.open(nonexistent_file, 'r')
-
-    def test_s3_open_write_does_not_raise_error_for_nonexistent_remote_file(self):
-        nonexistent_file = self.remote_file(str(uuid.uuid4()))
-        self.assert_s3_does_not_exist(nonexistent_file)
-        with s3.open(nonexistent_file, 'w') as f:
-            test_string = 'FOO!\n'
-            f.write(test_string)
-        self.assert_s3_exists(nonexistent_file)
-
-    def test_s3_open_write_update_raises_error_for_nonexistent_remote_file(self):
-        nonexistent_file = self.remote_file(str(uuid.uuid4()))
-        self.assert_s3_does_not_exist(nonexistent_file)
-        with self.assertRaises(NotImplementedError):
-            s3.open(nonexistent_file, 'w+')
-
-    def test_s3_open_exclusive_write_does_not_raise_error_for_nonexistent_remote_file(self):
-        nonexistent_file = self.remote_file(str(uuid.uuid4()))
-        self.assert_s3_does_not_exist(nonexistent_file)
-        with s3.open(nonexistent_file, 'x') as f:
-            test_string = 'FOO!\n'
-            f.write(test_string)
-        self.assert_s3_exists(nonexistent_file)
-
-    def test_s3_open_read_raises_error_for_nonexistent_local_file(self):
-        nonexistent_file = os.path.join(self.tmp_dir, "test_s3_open_read_raises_error_for_nonexistent_local_file")
-        self.assert_s3_does_not_exist(nonexistent_file)
-        with self.assertRaises(s3.KeyNotFound):
-            s3.open(nonexistent_file, 'r')
-
-    def test_s3_open_write_does_not_raise_error_for_nonexistent_local_file(self):
-        nonexistent_file = os.path.join(self.tmp_dir, "test_s3_open_write_does_not_raise_error_for_nonexistent_local_file")
-        self.assert_s3_does_not_exist(nonexistent_file)
-        with s3.open(nonexistent_file, 'w') as f:
-            test_string = 'FOO!\n'
-            f.write(test_string)
-        self.assert_s3_exists(nonexistent_file)
-
-    def test_s3_open_write_update_does_not_raise_error_for_nonexistent_local_file(self):
-        nonexistent_file = os.path.join(self.tmp_dir, "test_s3_open_write_update_does_not_raise_error_for_nonexistent_local_file")
-        self.assert_s3_does_not_exist(nonexistent_file)
-        with s3.open(nonexistent_file, 'w+') as f:
-            test_string = 'FOO!\n'
-            f.write(test_string)
-            f.seek(0)
-            self.assertEqual(test_string, f.read())
-        self.assert_s3_exists(nonexistent_file)
-
-    def test_s3_open_exclusive_write_does_not_raise_error_for_nonexistent_local_file(self):
-        nonexistent_file = os.path.join(self.tmp_dir, "test_s3_open_exclusive_write_does_not_raise_error_for_nonexistent_local_file")
-        self.assert_s3_does_not_exist(nonexistent_file)
-        with s3.open(nonexistent_file, 'x') as f:
-            test_string = 'FOO!\n'
-            f.write(test_string)
-        self.assert_s3_exists(nonexistent_file)
-
-    def test_s3_open_local_underlying_error_raises_ioerror_with_errno_and_strerror(self):
-        import errno
-        with self.assertRaises(IOError) as ctx:
-            s3.open(os.getcwd(), 'w')
-        self.assertEquals(ctx.exception.errno, errno.EISDIR)
-        self.assertIn('Is a directory:', ctx.exception.strerror)
-
-    @mock.patch('baiji.s3.CachedFile.upload')
-    def test_s3_open_write_calls_upload(self, upload):
-        remote_file_name = self.remote_file("write_test_1")
-        with s3.open(remote_file_name, 'w') as f:
-            self.assertFalse(upload.called)
-        self.assertTrue(upload.called)
-
-    @mock.patch('baiji.s3.CachedFile.upload')
-    def test_s3_open_write_does_not_upload_if_exception_raised(self, upload):
-        remote_file_name = self.remote_file("write_test_1")
-        try:
-            with s3.open(remote_file_name, 'w') as f:
-                raise AttributeError()
-        except AttributeError:
-            pass
-        self.assertFalse(upload.called)
-
-        # Sanity check
-        with s3.open(remote_file_name, 'w') as f:
-            pass
-        self.assertTrue(upload.called)
-
-    @mock.patch('baiji.s3.CachedFile.upload')
-    def test_s3_open_read_does_not_call_upload(self, upload):
-        self.assert_s3_exists(self.remote_file("openable"))
-        with s3.open(self.remote_file("openable"), 'r') as f:
-            pass
-        self.assertFalse(upload.called)
-
 class TestS3ConnectionPersistence(unittest.TestCase):
     def test_connection_is_cached(self):
+        # pylint: disable=protected-access
         # Using a cached connection is 2 orders of magnitude faster than
         # creating a new connection instance.
